Add files via upload
Creates a BruinBus shapefile that displays the routes, and contains trips, stops, and stoptimes joined together in the attribute table. Regarding the FolderPath, it uses the "MyProject26" folder that was shared, however within that folder my code uses a folder I called "BruinBusGTFS", beyond that everything joins to the absolute FolderPath. If you create a "BruinBusGTFS" folder within the "MyProject26" folder on your computer that contains the raw Bruin Bus gtfs text data, I think this should run for you.